Handle help, version and @file parameters in scalac and scaladoc
