+++ compiler/test/dotty/tools/dotc/SettingsTests.scala
@@ -47,7 +47,7 @@ class SettingsTests {
 
     inContext {
       val args = List("-foo", "b", "-bar", "1")
-      val summary = Settings.processArguments(args, true)
+      val summary = Settings.processArguments(args, summon[Context].settingsState, true)
       assertTrue(summary.errors.isEmpty)
       assertEquals("b", Settings.foo.value)
       assertEquals(1, Settings.bar.value)
@@ -64,7 +64,7 @@ class SettingsTests {
 
     inContext {
       val args = List("-foo", "b", "-bar", "1", "-baz", "5")
-      val summary = Settings.processArguments(args, true)
+      val summary = Settings.processArguments(args, summon[Context].settingsState, true)
       assertTrue(summary.errors.isEmpty)
       assertEquals("b", Settings.foo.value)
       assertEquals(1, Settings.bar.value)
@@ -73,14 +73,14 @@ class SettingsTests {
 
     inContext {
       val args = List("-foo:b")
-      val summary = Settings.processArguments(args, true)
+      val summary = Settings.processArguments(args, summon[Context].settingsState, true)
       assertTrue(summary.errors.isEmpty)
       assertEquals("b", Settings.foo.value)
     }
 
     inContext {
       val args = List("-foo", "c", "-bar", "3", "-baz", "-1")
-      val summary = Settings.processArguments(args, true)
+      val summary = Settings.processArguments(args, summon[Context].settingsState, true)
       val expectedErrors = List(
         "c is not a valid choice for -foo",
         "3 is not a valid choice for -bar",
@@ -91,14 +91,14 @@ class SettingsTests {
 
     inContext {
       val args = List("-foo:c")
-      val summary = Settings.processArguments(args, true)
+      val summary = Settings.processArguments(args, summon[Context].settingsState, true)
       val expectedErrors = List("c is not a valid choice for -foo")
       assertEquals(expectedErrors, summary.errors)
     }
 
     inContext {
       val args = List("-quux", "a", "-quuz", "0")
-      val summary = Settings.processArguments(args, true)
+      val summary = Settings.processArguments(args, summon[Context].settingsState, true)
       val expectedErrors = List(
         "a is not a valid choice for -quux",
         "0 is not a valid choice for -quuz",
